Added basic account creation
Once logged in, the user enters the account management screen. From
their the user can choose to create a new account, which will then be
added to their list of accounts and also to the management screen. New
accounts are at the moment only initialized with a name and a starting
balance.